backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-genai==1.0.0
pydantic==2.9.0
python-dotenv==1.0.1
aiosmtplib==3.0.1
